Edited a few data types for better run-time efficiency and set movement checker

+++ Assets/Scripts/Map/cTile/Tile.cs
@@ -20,7 +20,7 @@ using UnityEngine;
 public class Tile: MonoBehaviour
 {
     public bool m_isBlockedFlying, m_isBlockedGround, m_isWater,  m_isSink, m_isDecorable;
-    public double m_damage;
+    public int m_damage;
     public int m_spriteID;
     public Decoration m_decoration { get; set; }
